Create getTasksGroupedByDate
This was done to share the tasks fetch and group with week and month
view
