+++ frontend/src/app/tasks/week/page.tsx
@@ -1,43 +1,16 @@
-import { makeGetTasks } from '@/infra/task/getTasks'
-import { serverFetch } from '@/infra/lib/serverFetch'
-import { authOptions } from '@/app/api/auth/[...nextauth]/route'
-import { getServerSession } from 'next-auth'
 import { format, eachDayOfInterval, startOfWeek, add, isSameDay } from 'date-fns'
-import { Task } from '@/domain/Task'
 import { Typography, Box, Stack } from '@mui/joy'
 import { SimpleTaskBox } from '../components/TaskBox'
-import { convertTimeToMinutes, convertMinutesToTime } from '../utils/time'
+import { convertMinutesToTime } from '../utils/time'
 
-const getTasks = async (startTime: string, endTime: string) => {
-  const apiClient = await serverFetch()
-  const getTasks = makeGetTasks(apiClient)
-  const session = await getServerSession(authOptions)
-  const { id: userId } = session!.user
-
-  const tasks = await getTasks({ userId, startTime, endTime })
-
-  // Group tasks by date and add the total time for that date
-  return tasks.reduce<Record<Task['date'], { tasks: Array<Task>; time: number }>>((acc, task) => {
-    const startTime = convertTimeToMinutes(task.startTime)
-    const endTime = convertTimeToMinutes(task.endTime)
-    const timeDiff = endTime - startTime
-
-    if (acc[task.date]) {
-      const { tasks, time } = acc[task.date]
-
-      return { ...acc, [task.date]: { tasks: [...tasks, task], time: time + timeDiff } }
-    }
-
-    return { ...acc, [task.date]: { tasks: [task], time: timeDiff } }
-  }, {})
-}
+import { getTasksGroupedByDate } from '../actions/getTasksGroupedByDate'
 
 export default async function WeekView() {
   const today = new Date()
   const firstDayofWeek = startOfWeek(today)
   const lastDayOfWeek = add(firstDayofWeek, { days: 6 })
   const dateRange = eachDayOfInterval({ start: firstDayofWeek, end: lastDayOfWeek })
-  const groupedTasks = await getTasks(
+  const groupedTasks = await getTasksGroupedByDate(
     format(firstDayofWeek, 'yyyy-MM-dd'),
     format(lastDayOfWeek, 'yyyy-MM-dd')
   )
